# doc/weather.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Config Setup Service",
    "version": "1.0",
    "contact": {
      "name": "gRPC-Config Setup Service",
      "url": "https://github.com/khusainnov/sbercloud",
      "email": "khusainnov@gmail.com"
    }
  },
  "tags": [
    {
      "name": "UploadConfigService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/config": {
      "post": {
        "summary": "Adding new version of config",
        "description": "Uploading .json config file",
        "operationId": "UploadConfigService_UploadConfig",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ConfigBodyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Config"
            }
          }
        ],
        "tags": [
          "UploadConfigService"
        ]
      }
    },
    "/v1/config/{service}": {
      "get": {
        "summary": "Getting config body",
        "description": "Sending request with {config_name@version[0-9]}",
        "operationId": "UploadConfigService_GetConfig",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ConfigBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "UploadConfigService"
        ]
      },
      "delete": {
        "summary": "Delete config",
        "description": "Sending request with {config_name@version[0-9]}",
        "operationId": "UploadConfigService_DeleteConfig",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ConfigBodyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "UploadConfigService"
        ]
      },
      "patch": {
        "summary": "Update exists version of config",
        "description": "Updating .json config file",
        "operationId": "UploadConfigService_UpdateConfig",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ConfigBodyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "byte"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "tags": [
          "UploadConfigService"
        ]
      }
    }
  },
  "definitions": {
    "Config": {
      "type": "object",
      "properties": {
        "service": {
          "type": "string",
          "format": "byte"
        },
        "data": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "ConfigBody": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "ConfigBodyResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}